cleaned up test cases
